+++ tests/test_lif_plate_extended.py
@@ -2,10 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from fractal_lif_converters.common import image_in_plate_compute_task
-from fractal_lif_converters.lif_plate.convert_lif_plate_init_task import (
-    convert_lif_plate_init_task,
-)
+from fractal_lif_converters import convert_lif_plate
 
 from .utils import run_converter_test
 
@@ -75,7 +72,7 @@ _XLEF_TIF_DATASETS: list[str] = [
     [
         pytest.param(
             {"acquisitions": [{"path": str(LIF_RAW_DIR / f"{name}.lif")}]},
-            LIF_SNAPSHOT_DIR / f"{name}_plate.yaml",
+            LIF_SNAPSHOT_DIR / f"{name}.yaml",
             id=name,
         )
         for name in _DATASETS
@@ -87,7 +84,7 @@ _XLEF_TIF_DATASETS: list[str] = [
                     {"path": str(XLEF_RAW_DIR / f"{name}" / f"{name}.xlef")}
                 ]
             },
-            XLEF_SNAPSHOT_DIR / f"{name}_plate.yaml",
+            XLEF_SNAPSHOT_DIR / f"{name}.yaml",
             id=name,
         )
         for name in _XLEF_DATASETS
@@ -99,7 +96,7 @@ _XLEF_TIF_DATASETS: list[str] = [
                     {"path": str(XLEF_TIF_RAW_DIR / f"{name}" / f"{name}.xlef")}
                 ]
             },
-            XLEF_TIF_SNAPSHOT_DIR / f"{name}_plate.yaml",
+            XLEF_TIF_SNAPSHOT_DIR / f"{name}.yaml",
             id=name,
         )
         for name in _XLEF_TIF_DATASETS
@@ -114,9 +111,8 @@ def test_lif_plate_extended(
 ):
     run_converter_test(
         tmp_path=tmp_path,
-        init_task_fn=convert_lif_plate_init_task,
-        compute_task_fn=image_in_plate_compute_task,
-        init_task_kwargs=init_task_kwargs,
+        api_fn=convert_lif_plate,
+        api_kwargs=init_task_kwargs,
         snapshot_path=snapshot_path,
         update_snapshots=update_snapshots,
         converter_options=converter_options,